@@ -20,70 +20,83 @@
   "license": "MIT",
   "dependencies": {
     "algoliasearch": "^3.9.2",
-    "alt": "^0.17.1",
+    "alt": "^0.18.1",
     "alt-container": "^1.0.0",
-    "axios": "github:iam4x/axios#master",
+    "alt-utils": "^1.0.0",
+    "autoprefixer": "^6.1.0",
+    "axios": "^0.8.1",
     "babel": "^5.8.3",
+    "babel-core": "^5.8.34",
+    "babel-loader": "^5.4.0",
+    "babel-runtime": "^5.8.34",
+    "bird3-purifycss-webpack-plugin": "^1.0.9",
     "blocked": "^1.1.0",
     "can-use-dom": "^0.1.0",
     "classnames": "^2.1.3",
     "connect-alt": "^1.0.0",
     "cookies-js": "^1.2.1",
+    "css-loader": "^0.23.0",
     "debug": "^2.2.0",
+    "enquire.js": "git://github.com/adamjv90/enquire.js.git",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "file-loader": "^0.8.4",
     "font-awesome": "^4.4.0",
-    "history": "^1.9.1",
+    "history": "^1.17.0",
+    "horsey": "^3.0.0",
+    "image-webpack-loader": "^1.6.2",
     "immutable": "^3.7.5",
     "intl": "^1.0.0",
-    "iso": "^4.1.0",
-    "koa": "^1.0.0",
-    "koa-cash": "^1.0.1",
+    "iso": "github:iam4x/iso#master",
+    "json-loader": "^0.5.3",
+    "json2mq": "^0.2.0",
+    "koa": "^2.0.0-alpha.3",
     "koa-compressor": "^1.0.3",
     "koa-conditional-get": "^1.0.3",
+    "koa-convert": "^1.2.0",
     "koa-etag": "^2.0.0",
     "koa-favicon": "^1.2.0",
-    "koa-hbs": "^0.7.0",
-    "koa-helmet": "^0.3.0",
-    "koa-logger": "^1.3.0",
+    "koa-helmet": "^2.0.0-alpha.1",
+    "koa-logger": "^2.0.0",
     "koa-mount": "^1.3.0",
     "koa-proxy": "^0.4.0",
     "koa-response-time": "^1.0.2",
-    "koa-router": "^5.2.3",
+    "koa-router": "^7.0.0",
     "koa-static-cache": "^3.1.2",
     "lodash": "^3.10.1",
     "normalize.css": "^3.0.3",
+    "postcss-import": "^7.1.3",
+    "postcss-loader": "^0.8.0",
+    "postcss-url": "^5.0.2",
+    "precss": "^1.3.0",
     "react": "^0.14.0",
+    "react-addons-clone-with-props": "^0.14.6",
     "react-addons-linked-state-mixin": "^0.14.2",
     "react-addons-update": "^0.14.3",
     "react-dom": "^0.14.0",
     "react-element-pan": "^1.0.4",
+    "react-helmet": "^2.2.0",
     "react-imageloader": "^2.1.0",
-    "react-intl": "^1.2.0",
+    "react-intl": "1.2.0",
     "react-router": "^1.0.0",
+    "scroll-behavior": "^0.3.0",
     "shortid": "^2.2.2",
-    "velocity-animate": "^1.2.3"
+    "striptags": "^2.1.1",
+    "style-loader": "^0.13.0",
+    "velocity-animate": "^1.2.3",
+    "webpack": "^1.12.6"
   },
   "devDependencies": {
-    "autoprefixer": "^6.0.3",
-    "babel-core": "^5.8.3",
     "babel-eslint": "^4.0.5",
-    "babel-loader": "^5.3.2",
     "babel-plugin-react-transform": "^1.0.3",
-    "babel-runtime": "^5.8.24",
-    "bird3-purifycss-webpack-plugin": "github:iam4x/bird3-purifycss-webpack-plugin#master",
+    "blocked": "^1.1.0",
     "browser-sync": "^2.8.0",
-    "css-loader": "^0.22.0",
-    "del": "^2.0.2",
     "dev-ip": "^1.0.1",
     "eslint": "^1.4.1",
-    "eslint-config-airbnb": "^1.0.0",
+    "eslint-config-airbnb": "^3.0.0",
     "eslint-loader": "^1.0.0",
     "eslint-plugin-react": "^3.0.0",
-    "extract-text-webpack-plugin": "^0.9.0",
     "faux-jax": "^4.1.0",
-    "file-loader": "^0.8.4",
-    "image-webpack-loader": "^1.6.1",
     "isparta-loader": "^1.0.0",
-    "json-loader": "^0.5.2",
     "karma": "^0.13.3",
     "karma-chai-plugins": "^0.6.0",
     "karma-cli": "^0.1.0",
