@@ -1,45 +1,35 @@
-import React, { Component, PropTypes } from 'react';
+import React, { PropTypes, Children, cloneElement } from 'react';
+import classNames from 'classnames';
 import Modal from 'components/modal';
+import ResponsiveMixin from 'mixins/responsive';
 
 if (process.env.BROWSER) require('styles/app.css');
 
-class App extends Component {
+export default React.createClass({
+  mixins: [ ResponsiveMixin ],
 
-  static propTypes = {
+  propTypes: {
     children: PropTypes.element,
     location: PropTypes.object
-  }
+  },
 
-  static contextTypes = {
+  contextTypes: {
     flux: PropTypes.object.isRequired
-  }
+  },
 
-  static childContextTypes = {
-    messages: PropTypes.object.isRequired,
-    locales: PropTypes.array.isRequired
-  }
-
-  state = { i18n: this.context
-      .flux.getStore('locale').getState() }
-
-  getChildContext() {
-    const { i18n: { messages, locales } } = this.state;
-    return { messages, locales };
-  }
+  getInitialState() {
+    return { i18n: this.context.flux.getStore('locale').getState() };
+  },
 
   componentDidMount() {
     const { flux } = this.context;
-
-    flux.getStore('locale').listen(this.handleLocaleChange);
-    flux.getStore('title').listen(this.handleTitleChange);
-  }
+    flux.getStore('helmet').listen(this.handleTitleChange);
+  },
 
   componentWillUnmount() {
     const { flux } = this.context;
-
-    flux.getStore('locale').unlisten(this.handleLocaleChange);
-    flux.getStore('title').unlisten(this.handleTitleChange);
-  }
+    flux.getStore('helmet').unlisten(this.handleTitleChange);
+  },
 
   componentWillReceiveProps() {
     const { location } = this.props;
@@ -51,10 +41,21 @@ class App extends Component {
     if (!isModal) {
       this.previousChildren = this.props.children;
     }
-  }
+  },
 
-  handleLocaleChange = (i18n) => this.setState({ i18n })
-  handleTitleChange = ({ title }) => document.title = title
+  handleLocaleChange(i18n) {
+    this.setState({ i18n });
+  },
+
+  handleTitleChange({ titleBase, title }) {
+    document.title = titleBase + title;
+  },
+
+  cloneWithDimensions(children) {
+    return Children.map(children, (child) => {
+      return cloneElement(child, this.state);
+    });
+  },
 
   render() {
     const { location } = this.props;
@@ -64,20 +65,18 @@ class App extends Component {
       this.previousChildren
     );
     return (
-      <div style={ { height: '100%' } }>
+      <div style={ { height: '100%' } } className={ classNames({ tablet: this.state.isTablet, mobile: this.state.isMobile }) }>
         { isModal ?
-          this.previousChildren :
-          this.props.children
+          this.cloneWithDimensions(this.previousChildren) :
+          this.cloneWithDimensions(this.props.children)
         }
         { isModal && (
           <Modal returnTo={ location.state.returnTo } location={ location }>
-            { this.props.children }
+            { this.cloneWithDimensions(this.props.children) }
           </Modal>
         ) }
       </div>
     );
   }
 
-}
-
-export default App;
+});
