@@ -1,4 +1,5 @@
 export { default as users } from './users';
 export { default as locale } from './locale';
-export { default as title } from './page-title';
+export { default as helmet } from './helmet';
 export { default as requests } from './requests';
+export { default as session } from './session';
