@@ -1,5 +1,5 @@
 import debug from 'debug';
-import createBrowserHistory from 'history/lib/createBrowserHistory';
+import history from 'utils/router-history';
 
 // Paths are relative to `app` directory
 import createFlux from 'flux/createFlux';
@@ -12,7 +12,6 @@ if (NODE_ENV === 'development') debug.enable('dev,koa');
 
 const client = new ApiClient();
 const flux = createFlux(client);
-const history = createBrowserHistory();
 
 universalRender({ flux, history })
   .catch(err => debug('dev')(err));
