@@ -4,19 +4,19 @@ const simplifyUsers = (collection) => collection
   .map(({ user, seed }) => ({ ...user, seed }))
   .map(({ email, name, seed, picture }) => ({ email, name, seed, picture }));
 
-export default function(router) {
-  router.get('/users', function *() {
-    this.body = simplifyUsers(users.slice(0, 10));
+export default function (router) {
+  router.get('/users', async function (ctx) {
+    ctx.body = simplifyUsers(users.slice(0, 10));
   });
 
-  router.get('/users/:seed', function *() {
-    const { seed } = this.params;
+  router.get('/users/:seed', async function (ctx) {
+    const { seed } = ctx.params;
     const [ result ] = simplifyUsers(users.filter(user => user.seed === seed));
 
     if (!result) {
-      this.body = { error: { message: 'User not found' } };
+      ctx.body = { error: { message: 'User not found' } };
     } else {
-      this.body = result;
+      ctx.body = result;
     }
   });
 }
