@@ -1,8 +1,7 @@
 import path from 'path';
 import debug from 'debug';
 
-import koa from 'koa';
-import hbs from 'koa-hbs';
+import Koa from 'koa';
 import mount from 'koa-mount';
 import helmet from 'koa-helmet';
 import logger from 'koa-logger';
@@ -10,16 +9,17 @@ import favicon from 'koa-favicon';
 import staticCache from 'koa-static-cache';
 import responseTime from 'koa-response-time';
 import Router from 'koa-router';
+import convert from 'koa-convert';
 
 import router from './router';
 import config from './config/init';
 
-const app = koa();
+const app = new Koa();
 const env = process.env.NODE_ENV || 'development';
 
 // add header `X-Response-Time`
-app.use(responseTime());
-app.use(logger());
+app.use(convert(responseTime()));
+app.use(convert(logger()));
 
 // various security headers
 app.use(helmet());
@@ -30,9 +30,9 @@ if (env === 'production') {
   debug.enable('koa');
 
   // load production middleware
-  app.use(require('koa-conditional-get')());
-  app.use(require('koa-etag')());
-  app.use(require('koa-compressor')());
+  app.use(convert(require('koa-conditional-get')()));
+  app.use(convert(require('koa-etag')()));
+  app.use(convert(require('koa-compressor')()));
 }
 
 if (env === 'development') {
@@ -42,12 +42,7 @@ if (env === 'development') {
   require('blocked')((ms) => debug('koa')(`blocked for ${ms}ms`));
 }
 
-app.use(favicon(path.join(__dirname, '../app/images/favicon.ico')));
-app.use(hbs.middleware({
-  defaultLayout: 'index',
-  layoutsPath: path.join(__dirname, '/views/layouts'),
-  viewPath: path.join(__dirname, '/views')
-}));
+app.use(convert(favicon(path.join(__dirname, '../app/images/favicon.ico'))));
 
 const cacheOpts = { maxAge: 86400000, gzip: true };
 
@@ -58,9 +53,9 @@ if (env === 'development') {
     host: 'http://0.0.0.0:' + webpackConfig.server.port,
     map: (filePath) => 'assets/' + filePath
   });
-  app.use(mount('/assets', proxy));
+  app.use(convert(mount('/assets', proxy)));
 } else {
-  app.use(mount('/assets', staticCache(path.join(__dirname, '../dist'), cacheOpts)));
+  app.use(convert(mount('/assets', staticCache(path.join(__dirname, '../dist'), cacheOpts))));
 }
 
 // mount `/api` router
