@@ -19,14 +19,14 @@ const runRouter = (location, routes) =>
 
 const bootstrap = () =>
   new Promise((resolve) =>
-    Iso.bootstrap((initialState, __, container) =>
-      resolve({ initialState, __, container })));
+    Iso.bootstrap((initialState, container) =>
+      resolve({ initialState, container })));
 
 /* eslint space-before-function-paren:0 */
 // https://github.com/eslint/eslint/issues/4442
 export default async function({ flux, history, location }) {
   if (BROWSER) {
-    if (NODE_ENV === 'development') require('alt/utils/chromeDebug')(flux);
+    if (NODE_ENV === 'development') require('alt-utils/lib/chromeDebug')(flux);
 
     const { container, initialState } = await bootstrap();
     flux.bootstrap(initialState);
@@ -35,15 +35,18 @@ export default async function({ flux, history, location }) {
     // load the correct data/{lang}.json into app
     const { locales: [ locale ] } = flux.getStore('locale').getState();
     const { messages } = await intlLoader(locale);
-    flux.getActions('locale').switchLocaleSuccess({ locale, messages });
+    flux.getActions('locale').switchLocale({ locale, messages });
 
     const routes = require('routes');
+    const I18nContainer = require('utils/i18n-container');
 
     const element = (
       <AltContainer flux={ flux }>
-        <Router
-          history={ history }
-          routes={ routes } />
+        <I18nContainer>
+          <Router
+            history={ history }
+            routes={ routes(flux) } />
+        </I18nContainer>
       </AltContainer>
     );
 
@@ -54,14 +57,17 @@ export default async function({ flux, history, location }) {
     // next promises will be resolved
     flux.resolver.firstRender = false;
   } else {
-    const routes = require('routes');
+    const routes = require('routes')(flux);
+    const I18nContainer = require('utils/i18n-container');
     const [ error, redirect, renderProps ] = await runRouter(location, routes);
 
     if (error || redirect) throw ({ error, redirect });
 
     const element = (
       <AltContainer flux={ flux }>
-        <RoutingContext { ...renderProps } />
+        <I18nContainer>
+          <RoutingContext { ...renderProps } />
+        </I18nContainer>
       </AltContainer>
     );
 
@@ -88,7 +94,13 @@ export default async function({ flux, history, location }) {
       app = renderToString(<ErrorPage />);
     }
 
-    const { title } = flux.getStore('title').getState();
-    return { body: Iso.render(app, fluxSnapshot), title };
+    // Get status code, page title and page description for rendering
+    const { titleBase, title, ...helmet } = flux.getStore('helmet').getState();
+
+    return {
+      ...helmet,
+      body: Iso.render(app, fluxSnapshot),
+      title: titleBase + title
+    };
   }
 }
