@@ -26,16 +26,16 @@ export default {
   },
   plugins: [
     // write webpack stats
-    function() { this.plugin('done', writeStats); }
+    function () { this.plugin('done', writeStats); }
   ],
   resolve: {
-    extensions: ['', '.js', '.json', '.jsx', '.es6', '.babel'],
-    modulesDirectories: ['node_modules', 'app']
+    extensions: [ '', '.js', '.json', '.jsx', '.es6', '.babel' ],
+    modulesDirectories: [ 'node_modules', 'app' ]
   },
   postcss: (webpack) => [
     require('postcss-import')({ addDependencyTo: webpack }),
     require('postcss-url')(),
     require('precss')(),
-    require('autoprefixer')({ browsers: ['last 2 versions'] })
+    require('autoprefixer')({ browsers: [ 'last 2 versions' ] })
   ]
 };
