@@ -1,20 +1,21 @@
 process.env.BABEL_ENV = 'browser';
 process.env.NODE_ENV = 'development';
 
-import koa from 'koa';
+import Koa from 'koa';
 import debug from 'debug';
 import webpack from 'webpack';
+import convert from 'koa-convert';
 
 import config from './dev.config';
 
-const app = koa();
+const app = new Koa();
 const compiler = webpack(config.webpack);
 
 debug.enable('dev');
 
-app.use(require('koa-webpack-dev-middleware')(compiler, config.server.options));
-app.use(require('koa-webpack-hot-middleware')(compiler));
+app.use(convert(require('koa-webpack-dev-middleware')(compiler, config.server.options)));
+app.use(convert(require('koa-webpack-hot-middleware')(compiler)));
 
-app.listen(config.server.port, '0.0.0.0', function() {
+app.listen(config.server.port, '0.0.0.0', function () {
   debug('dev')('`webpack-dev-server` listening on port %s', config.server.port);
 });
