@@ -1,16 +1,14 @@
+/* eslint max-len: 0 */
+
 import webpack from 'webpack';
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import PurifyCSSPlugin from 'bird3-purifycss-webpack-plugin';
 
 import baseConfig from './base.config';
 
-// clean `.tmp` && `dist`
-require('./utils/clean-dist')();
-
 export default {
   ...baseConfig,
   module: {
-    ...baseConfig.module,
     loaders: [
       ...baseConfig.module.loaders,
       {
