@@ -5,17 +5,17 @@ import debug from 'debug';
 
 const filepath = path.resolve(__dirname, '../../server/webpack-stats.json');
 
-export default function(stats) {
+export default function (stats) {
   const publicPath = this.options.output.publicPath;
   const json = stats.toJson();
 
   // get chunks by name and extensions
-  const getChunks = function(name, ext = /.js$/) {
+  const getChunks = function (name, ext = /.js$/) {
     let chunks = json.assetsByChunkName[name];
 
     // a chunk could be a string or an array, so make sure it is an array
     if (!(Array.isArray(chunks))) {
-      chunks = [chunks];
+      chunks = [ chunks ];
     }
 
     return chunks
